backend/services/scheduler.py
import logging
from apscheduler.schedulers.asyncio import AsyncIOScheduler
from apscheduler.triggers.cron import CronTrigger
from datetime import datetime
import os
from ..services.telegram import TelegramService
import asyncio

logger = logging.getLogger(__name__)

class TradingScheduler:
    def __init__(self, telegram_service: TelegramService = None):
        """Initialize the trading scheduler with optional telegram service"""
        self.scheduler = AsyncIOScheduler()
        self.telegram_service = telegram_service
        self._setup_jobs()

    def _setup_jobs(self):
        """Setup scheduled jobs"""
        # Daily summary at midnight
        self.scheduler.add_job(
            self._send_daily_summary,
            CronTrigger(hour=0, minute=0),
            id='daily_summary'
        )

        # Market analysis every hour
        self.scheduler.add_job(
            self._analyze_market,
            CronTrigger(minute=0),
            id='market_analysis'
        )

    async def _send_daily_summary(self):
        """Send daily trading summary"""
        try:
            if self.telegram_service:
                # Get daily summary data (implement this based on your needs)
                summary = {
                    "total_trades": 0,
                    "winning_trades": 0,
                    "losing_trades": 0,
                    "net_profit": 0.0,
                    "trades": []
                }
                await self.telegram_service.send_daily_summary(summary)
                logger.info("Daily summary sent successfully")
        except Exception as e:
            logger.error(f"Error sending daily summary: {str(e)}")

    async def _analyze_market(self):
        """Perform market analysis and send notifications if needed"""
        try:
            if self.telegram_service and self.telegram_service.market_analyzer:
                # Get market analysis for default pair
                symbol = os.getenv('DEFAULT_TRADING_PAIR', 'BTC/USDT')
                analysis = await self.telegram_service.market_analyzer.get_market_analysis(symbol)

                # Send notification if significant changes detected
                if self._should_notify(analysis):
                    await self.telegram_service.send_market_alert(symbol, analysis)
                    logger.info(f"Market alert sent for {symbol}")
        except Exception as e:
            logger.error(f"Error analyzing market: {str(e)}")

    def _should_notify(self, analysis: dict) -> bool:
        """Determine if a market alert should be sent based on analysis"""
        # Implement your notification criteria here
        return False  # Placeholder

    def start(self):
        """Start the scheduler"""
        try:
            self.scheduler.start()
            logger.info("Trading scheduler started")
        except Exception as e:
            logger.error(f"Error starting scheduler: {str(e)}")
            raise

    def stop(self):
        """Stop the scheduler"""
        try:
            self.scheduler.shutdown()
            logger.info("Trading scheduler stopped")
        except Exception as e:
            logger.error(f"Error stopping scheduler: {str(e)}")
            raise
