requirements.txt
accelerate==1.5.2
aiodns==3.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.4.0
ccxt==4.4.77
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
controlnet-aux==0.0.9
cryptography==44.0.2
cssselect==1.2.0
cycler==0.12.1
diffusers==0.32.2
ecdsa==0.19.0
einops==0.8.1
fake-useragent==1.5.1
fastapi==0.115.12
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
frozendict==2.4.4
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.23.1
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
h11==0.14.0
holidays==0.68
html5lib==1.1
httpcore==1.0.7
httpx==0.26.0
huggingface-hub==0.29.3
idna==3.7
imageio==2.37.0
importlib_metadata==8.0.0
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
narwhals==1.29.0
networkx==3.4.2
numpy==1.26.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.16
package_name==0.1
packaging==24.1
pandas==2.2.0
parse==1.20.2
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.18.0
pluggy==1.5.0
propcache==0.3.1
prophet==1.1.5
protobuf==5.27.2
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.1
pycares==4.6.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pyparsing==3.2.1
pypdfium2==4.30.1
pyppeteer==2.0.0
pyquery==2.0.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-html==0.10.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.4.0
scipy==1.14.0
semantic-version==2.10.0
setuptools==78.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.25
stanio==0.5.1
starlette==0.46.1
streamlit==1.36.0
streamlit-option-menu==0.3.13
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2025.3.13
timm==0.6.7
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.50.3
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.27.0
w3lib==2.2.1
watchdog==4.0.1
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
xgboost==2.1.4
yahoo-fin==0.8.9.1
yarl==1.20.0
yfinance==0.2.40
zipp==3.19.2
